Refractored all the endpoint
