+++ api-gateway/src/main/java/com/petclinic/bffapigateway/presentationlayer/BFFApiGatewayController.java
@@ -345,149 +345,149 @@ public class BFFApiGatewayController {
 
 */
 
-        /* Visits Methods */
-
-    @SecuredEndpoint(allowedRoles = {Roles.ADMIN})
-    @GetMapping(value = "reviews", produces = MediaType.TEXT_EVENT_STREAM_VALUE)
-    public Flux<ReviewResponseDTO> getAllReviews(){
-        return visitsServiceClient.getAllReviews();
-    }
-
-
-
-
-
-
-
-
-
-
-    @SecuredEndpoint(allowedRoles = {Roles.ADMIN})
-    @GetMapping(value = "visits", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
-    public Flux<VisitResponseDTO> getAllVisits(@RequestParam(required = false) String description){
-        return visitsServiceClient.getAllVisits(description);
-    }
-
-    @SecuredEndpoint(allowedRoles = {Roles.ADMIN,Roles.VET,Roles.OWNER})
-    @IsUserSpecific(idToMatch = {"ownerId"}, bypassRoles = {Roles.ADMIN,Roles.VET})
-    @GetMapping(value = "visits/owners/{ownerId}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
-    public Flux<VisitResponseDTO> getVisitsByOwnerId(@PathVariable String ownerId){
-//not ideal since returns complete pet dto
-        return getPetsByOwnerId(ownerId).flatMap(petResponseDTO -> getVisitsForPet(petResponseDTO.getPetId()));
-    }
-
-
-    @SecuredEndpoint(allowedRoles = {Roles.ADMIN,Roles.VET,Roles.OWNER})
-    @IsUserSpecific(idToMatch = {"ownerId"}, bypassRoles = {Roles.ADMIN,Roles.VET})
-    @GetMapping(value = "visits/emergency/owners/{ownerId}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
-    public Flux<EmergencyResponseDTO> getEmergencyVisitsByOwnerId(@PathVariable String ownerId){
-//not ideal since returns complete pet dto
-        return getPetsByOwnerId(ownerId).flatMap(petResponseDTO -> getEmergencyVisitsForPet(petResponseDTO.getPetId()));
-    }
-
-    @GetMapping(value = "visits/vets/{practitionerId}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
-    public Flux<VisitResponseDTO> getVisitByPractitionerId(@PathVariable String practitionerId){
-        return visitsServiceClient.getVisitByPractitionerId(practitionerId);
-    }
-
-    @GetMapping(value = "visits/pets/{petId}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
-    public Flux<VisitResponseDTO> getVisitsForPet(@PathVariable String petId){
-        return visitsServiceClient.getVisitsForPet(petId);
-    }
-
-    @GetMapping(value = "visits/emergency/pets/{petId}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
-    public Flux<EmergencyResponseDTO> getEmergencyVisitsForPet(@PathVariable String petId){
-        return visitsServiceClient.getEmergencyVisitForPet(petId);
-    }
-
-    @GetMapping(value = "visits/status/{status}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
-    public Flux<VisitResponseDTO> getVisitsForStatus(@PathVariable String status){
-        return visitsServiceClient.getVisitsForStatus(status);
-    }
-
-    @GetMapping(value ="visits/{visitId}")
-    public Mono<VisitResponseDTO> getVisitByVisitId(@PathVariable String visitId){
-        return visitsServiceClient.getVisitByVisitId(visitId);
-    }
-    @PostMapping(value = "visit/owners/{ownerId}/pets/{petId}/visits", consumes = "application/json", produces = "application/json")
-    Mono<ResponseEntity<VisitResponseDTO>> addVisit(@RequestBody VisitRequestDTO visit, @PathVariable String ownerId, /*@PathVariable String petId,*/ @CookieValue("Bearer") String auth) {
-        visit.setOwnerId(ownerId);
-        visit.setJwtToken(auth);
-        return visitsServiceClient.createVisitForPet(visit).map(ResponseEntity.status(HttpStatus.CREATED)::body)
-                .defaultIfEmpty(ResponseEntity.badRequest().build());
-    }
-
-    @PutMapping(value = "/visits/{visitId}/status/{status}")
-    Mono<VisitResponseDTO> updateStatusForVisitByVisitId(@PathVariable String visitId, @PathVariable String status) {
-        return visitsServiceClient.updateStatusForVisitByVisitId(visitId, status);
-    }
-    @DeleteMapping (value = "visits/{visitId}")
-    public Mono<ResponseEntity<Void>> deleteVisitsByVisitId(@PathVariable String visitId){
-        return visitsServiceClient.deleteVisitByVisitId(visitId).then(Mono.just(ResponseEntity.noContent().<Void>build()))
-                .defaultIfEmpty(ResponseEntity.notFound().build());
-    }
-
-    @DeleteMapping(value = "visits/cancelled")
-    public Mono<ResponseEntity<Void>> deleteAllCancelledVisits(){
-        return visitsServiceClient.deleteAllCancelledVisits().then(Mono.just(ResponseEntity.noContent().<Void>build()))
-                .defaultIfEmpty(ResponseEntity.notFound().build());
-    }
-    //        @PutMapping(value = "owners/*/pets/{petId}/visits/{visitId}", consumes = "application/json", produces = "application/json")
-    /*
-        Mono<ResponseEntity<VisitDetails>> updateVisit(@RequestBody VisitDetails visit, @PathVariable String petId, @PathVariable String visitId) {
-            visit.setPetId(petId);
-            visit.setVisitId(visitId);
-            return visitsServiceClient.updateVisitForPet(visit).map(s -> ResponseEntity.status(HttpStatus.OK).body(s))
-                .defaultIfEmpty(ResponseEntity.badRequest().build());
-        }
-
-        @GetMapping(value = "visits/previous/{petId}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
-        public Flux<VisitDetails> getPreviousVisitsForPet(@PathVariable final String petId) {
-            return visitsServiceClient.getPreviousVisitsForPet(petId);
-        }
-        @GetMapping(value = "visits/scheduled/{petId}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
-        public Flux<VisitDetails> getScheduledVisitsForPet(@PathVariable final String petId) {
-            return visitsServiceClient.getScheduledVisitsForPet(petId);
-        }
-
-        @GetMapping(value = "visits/calendar/{practitionerId}")
-        public Flux<VisitDetails> getVisitsByPractitionerIdAndMonth(@PathVariable("practitionerId") int practitionerId, @RequestParam("dates") List<String> dates) {
-            String startDate = dates.get(0);
-            String endDate = dates.get(1);
-            return visitsServiceClient.getVisitsByPractitionerIdAndMonth(practitionerId, startDate, endDate);
-        }
-        private Function<Visits, OwnerResponseDTO> addVisitsToOwner(OwnerResponseDTO owner) {
-            return visits -> {
-                owner.getPets()
-                        .forEach(pet -> pet.getVisits()
-                                .addAll(visits.getItems().stream()
-                                        .filter(v -> v.getPetId() == pet.getId())
-                                        .collect(Collectors.toList()))
-                        );
-                return owner;
-            };
-        }
-*/
-
-    @PostMapping(value = "visit/owners/5fe81e29-1f1d-4f9d-b249-8d3e0cc0b7dd/pets/9/visits", consumes = "application/json", produces = "application/json")
-    Mono<ResponseEntity<VisitResponseDTO>> addVisit(@RequestBody VisitRequestDTO visit/* @PathVariable String ownerId, @PathVariable String petId*/) {
-       // visit.setPetId(petId);
-        return visitsServiceClient.createVisitForPet(visit).map(s -> ResponseEntity.status(HttpStatus.CREATED).body(s))
-                .defaultIfEmpty(ResponseEntity.badRequest().build());
-    }
-
-//    @PutMapping(value = "owners/*/pets/{petId}/visits/{visitId}", consumes = "application/json", produces = "application/json")
-//    Mono<ResponseEntity<VisitDetails>> updateVisit(@RequestBody VisitDetails visit, @PathVariable String petId, @PathVariable String visitId) {
-//        visit.setPetId(petId);
-//        visit.setVisitId(visitId);
-//        return visitsServiceClient.updateVisitForPet(visit).map(s -> ResponseEntity.status(HttpStatus.OK).body(s))
+//        /* Visits Methods */
+//
+//    @SecuredEndpoint(allowedRoles = {Roles.ADMIN})
+//    @GetMapping(value = "reviews", produces = MediaType.TEXT_EVENT_STREAM_VALUE)
+//    public Flux<ReviewResponseDTO> getAllReviews(){
+//        return visitsServiceClient.getAllReviews();
+//    }
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//    @SecuredEndpoint(allowedRoles = {Roles.ADMIN})
+//    @GetMapping(value = "visits", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
+//    public Flux<VisitResponseDTO> getAllVisits(@RequestParam(required = false) String description){
+//        return visitsServiceClient.getAllVisits(description);
+//    }
+//
+//    @SecuredEndpoint(allowedRoles = {Roles.ADMIN,Roles.VET,Roles.OWNER})
+//    @IsUserSpecific(idToMatch = {"ownerId"}, bypassRoles = {Roles.ADMIN,Roles.VET})
+//    @GetMapping(value = "visits/owners/{ownerId}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
+//    public Flux<VisitResponseDTO> getVisitsByOwnerId(@PathVariable String ownerId){
+////not ideal since returns complete pet dto
+//        return getPetsByOwnerId(ownerId).flatMap(petResponseDTO -> getVisitsForPet(petResponseDTO.getPetId()));
+//    }
+//
+//
+//    @SecuredEndpoint(allowedRoles = {Roles.ADMIN,Roles.VET,Roles.OWNER})
+//    @IsUserSpecific(idToMatch = {"ownerId"}, bypassRoles = {Roles.ADMIN,Roles.VET})
+//    @GetMapping(value = "visits/emergency/owners/{ownerId}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
+//    public Flux<EmergencyResponseDTO> getEmergencyVisitsByOwnerId(@PathVariable String ownerId){
+////not ideal since returns complete pet dto
+//        return getPetsByOwnerId(ownerId).flatMap(petResponseDTO -> getEmergencyVisitsForPet(petResponseDTO.getPetId()));
+//    }
+//
+//    @GetMapping(value = "visits/vets/{practitionerId}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
+//    public Flux<VisitResponseDTO> getVisitByPractitionerId(@PathVariable String practitionerId){
+//        return visitsServiceClient.getVisitByPractitionerId(practitionerId);
+//    }
+//
+//    @GetMapping(value = "visits/pets/{petId}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
+//    public Flux<VisitResponseDTO> getVisitsForPet(@PathVariable String petId){
+//        return visitsServiceClient.getVisitsForPet(petId);
+//    }
+//
+//    @GetMapping(value = "visits/emergency/pets/{petId}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
+//    public Flux<EmergencyResponseDTO> getEmergencyVisitsForPet(@PathVariable String petId){
+//        return visitsServiceClient.getEmergencyVisitForPet(petId);
+//    }
+//
+//    @GetMapping(value = "visits/status/{status}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
+//    public Flux<VisitResponseDTO> getVisitsForStatus(@PathVariable String status){
+//        return visitsServiceClient.getVisitsForStatus(status);
+//    }
+//
+//    @GetMapping(value ="visits/{visitId}")
+//    public Mono<VisitResponseDTO> getVisitByVisitId(@PathVariable String visitId){
+//        return visitsServiceClient.getVisitByVisitId(visitId);
+//    }
+//    @PostMapping(value = "visit/owners/{ownerId}/pets/{petId}/visits", consumes = "application/json", produces = "application/json")
+//    Mono<ResponseEntity<VisitResponseDTO>> addVisit(@RequestBody VisitRequestDTO visit, @PathVariable String ownerId, /*@PathVariable String petId,*/ @CookieValue("Bearer") String auth) {
+//        visit.setOwnerId(ownerId);
+//        visit.setJwtToken(auth);
+//        return visitsServiceClient.createVisitForPet(visit).map(ResponseEntity.status(HttpStatus.CREATED)::body)
 //                .defaultIfEmpty(ResponseEntity.badRequest().build());
 //    }
-    /*  End of Visit Methods */
-
-    /**
-     * End of Visit Methods
-     **/
+//
+//    @PutMapping(value = "/visits/{visitId}/status/{status}")
+//    Mono<VisitResponseDTO> updateStatusForVisitByVisitId(@PathVariable String visitId, @PathVariable String status) {
+//        return visitsServiceClient.updateStatusForVisitByVisitId(visitId, status);
+//    }
+//    @DeleteMapping (value = "visits/{visitId}")
+//    public Mono<ResponseEntity<Void>> deleteVisitsByVisitId(@PathVariable String visitId){
+//        return visitsServiceClient.deleteVisitByVisitId(visitId).then(Mono.just(ResponseEntity.noContent().<Void>build()))
+//                .defaultIfEmpty(ResponseEntity.notFound().build());
+//    }
+//
+//    @DeleteMapping(value = "visits/cancelled")
+//    public Mono<ResponseEntity<Void>> deleteAllCancelledVisits(){
+//        return visitsServiceClient.deleteAllCancelledVisits().then(Mono.just(ResponseEntity.noContent().<Void>build()))
+//                .defaultIfEmpty(ResponseEntity.notFound().build());
+//    }
+//    //        @PutMapping(value = "owners/*/pets/{petId}/visits/{visitId}", consumes = "application/json", produces = "application/json")
+//    /*
+//        Mono<ResponseEntity<VisitDetails>> updateVisit(@RequestBody VisitDetails visit, @PathVariable String petId, @PathVariable String visitId) {
+//            visit.setPetId(petId);
+//            visit.setVisitId(visitId);
+//            return visitsServiceClient.updateVisitForPet(visit).map(s -> ResponseEntity.status(HttpStatus.OK).body(s))
+//                .defaultIfEmpty(ResponseEntity.badRequest().build());
+//        }
+//
+//        @GetMapping(value = "visits/previous/{petId}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
+//        public Flux<VisitDetails> getPreviousVisitsForPet(@PathVariable final String petId) {
+//            return visitsServiceClient.getPreviousVisitsForPet(petId);
+//        }
+//        @GetMapping(value = "visits/scheduled/{petId}", produces= MediaType.TEXT_EVENT_STREAM_VALUE)
+//        public Flux<VisitDetails> getScheduledVisitsForPet(@PathVariable final String petId) {
+//            return visitsServiceClient.getScheduledVisitsForPet(petId);
+//        }
+//
+//        @GetMapping(value = "visits/calendar/{practitionerId}")
+//        public Flux<VisitDetails> getVisitsByPractitionerIdAndMonth(@PathVariable("practitionerId") int practitionerId, @RequestParam("dates") List<String> dates) {
+//            String startDate = dates.get(0);
+//            String endDate = dates.get(1);
+//            return visitsServiceClient.getVisitsByPractitionerIdAndMonth(practitionerId, startDate, endDate);
+//        }
+//        private Function<Visits, OwnerResponseDTO> addVisitsToOwner(OwnerResponseDTO owner) {
+//            return visits -> {
+//                owner.getPets()
+//                        .forEach(pet -> pet.getVisits()
+//                                .addAll(visits.getItems().stream()
+//                                        .filter(v -> v.getPetId() == pet.getId())
+//                                        .collect(Collectors.toList()))
+//                        );
+//                return owner;
+//            };
+//        }
+//*/
+//
+//    @PostMapping(value = "visit/owners/5fe81e29-1f1d-4f9d-b249-8d3e0cc0b7dd/pets/9/visits", consumes = "application/json", produces = "application/json")
+//    Mono<ResponseEntity<VisitResponseDTO>> addVisit(@RequestBody VisitRequestDTO visit/* @PathVariable String ownerId, @PathVariable String petId*/) {
+//       // visit.setPetId(petId);
+//        return visitsServiceClient.createVisitForPet(visit).map(s -> ResponseEntity.status(HttpStatus.CREATED).body(s))
+//                .defaultIfEmpty(ResponseEntity.badRequest().build());
+//    }
+//
+////    @PutMapping(value = "owners/*/pets/{petId}/visits/{visitId}", consumes = "application/json", produces = "application/json")
+////    Mono<ResponseEntity<VisitDetails>> updateVisit(@RequestBody VisitDetails visit, @PathVariable String petId, @PathVariable String visitId) {
+////        visit.setPetId(petId);
+////        visit.setVisitId(visitId);
+////        return visitsServiceClient.updateVisitForPet(visit).map(s -> ResponseEntity.status(HttpStatus.OK).body(s))
+////                .defaultIfEmpty(ResponseEntity.badRequest().build());
+////    }
+//    /*  End of Visit Methods */
+//
+//    /**
+//     * End of Visit Methods
+//     **/
 
 
     /**
