+++ petclinic-frontend/src/features/visits/Emergency/Api/getAllEmergency.ts
@@ -2,8 +2,8 @@ import axiosInstance from '@/shared/api/axiosInstance';
 import { EmergencyResponseDTO } from '../Model/EmergencyResponseDTO';
 
 export async function getAllEmergency(): Promise<EmergencyResponseDTO[]> {
-  const response = await axiosInstance.get(`/visits/emergency`, {
-    useV2: true,
+  const response = await axiosInstance.get(`/visits/emergencies`, {
+    useV2: false,
   });
   return response.data;
 }
