+++ petclinic-frontend/src/features/visits/Review/Api/getAllReviews.ts
@@ -4,7 +4,7 @@ import { ReviewResponseDTO } from '../Model/ReviewResponseDTO';
 export const getAllReviews = async (): Promise<ReviewResponseDTO[]> => {
   const response = await axiosInstance.get<ReviewResponseDTO[]>(
     `/visits/reviews`,
-    { useV2: true }
+    { useV2: false }
   );
   return response.data; // Return only the data
 };
