+++ petclinic-frontend/src/features/visits/api/getAllOwnerVisits.ts
@@ -5,9 +5,12 @@ export async function getAllOwnerVisits(
   ownerId: string
 ): Promise<VisitResponseModel[]> {
   try {
-    const response = await axiosInstance.get(`/visits/owners/${ownerId}`, {
-      useV2: true,
-    });
+    const response = await axiosInstance.get(
+      `/visits/owners/${ownerId}/visits`,
+      {
+        useV2: false,
+      }
+    );
     return response.data
       .split('data:')
       .map((dataChunk: string) => {
